requirements.txt
praw
anthropic
pyyaml
tabulate
httpx
